split `link_and_sew.rs` into  proper modules
